대미지팝업추가

+++ DungeonCleaner/Assets/Scripts/IDamagable.cs
@@ -2,5 +2,5 @@
 
 public interface IDamagable
 {
-    void OnDamage(float damage, Vector3 hitPoint, Vector3 hitNormal);
+    void OnDamage(int damage, Vector3 hitPoint, Vector3 hitNormal);
 }
